+++ test/authorizer/none-authorizer.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { getMockedRequest, mockConfig, mockFetch } from './authorizer-mocks.js';
+import { getMockedRequest, mockConfig, mockFetchUser } from './authorizer-mocks.js';
 
 vi.mock('../../src/config.js', () => {
   return mockConfig({
@@ -9,7 +9,7 @@ vi.mock('../../src/config.js', () => {
 
 vi.mock('node-fetch', () => {
   return {
-    default: mockFetch
+    default: mockFetchUser
   };
 });