+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-vi
 
 * Supported `AUTHORIZATION_SCHEME="fractal-server-viewer-paths"` (\#35);
 * Supported URL encoded characters in path (\#35);
+* Deprecated `AUTHORIZATION_SCHEME="allowed-list"` (\#35);
+* Removed support for relative paths (\#35);
 
 # 0.1.4