+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-viewer repository.
 
-# Unreleased
+# 0.2.0
 
 * Supported `AUTHORIZATION_SCHEME="fractal-server-viewer-paths"` (\#35);
 * Supported URL encoded characters in path (\#35);